+++ flask_admin/babel.py
@@ -4,11 +4,11 @@ try:
 except ImportError:
 
     def gettext(string, **variables):
-        return string % variables
+        return string if not variables else string % variables
 
     def ngettext(singular, plural, num, **variables):
         variables.setdefault("num", num)
-        return (singular if num == 1 else plural) % variables
+        return gettext((singular if num == 1 else plural), **variables)
 
     def lazy_gettext(string, **variables):
         return gettext(string, **variables)